+++ kernel/include/linux/fs/fat32.h
@@ -123,7 +123,7 @@
 	char *FAT32_get_shortname(int *namelen, const msdos_dirent_s *de);
 	u32 FAT32_read_FAT_Entry(FAT32_SBinfo_s * fsbi, u32 fat_entry);
 	u64 FAT32_write_FAT_Entry(FAT32_SBinfo_s * fsbi, u32 fat_entry, u32 value);
-	s64 FAT32_find_available_cluster(FAT32_SBinfo_s * fsbi);
+	u32 FAT32_find_available_cluster(FAT32_SBinfo_s * fsbi);
 	s64 FAT32_alloc_new_dir(inode_s *dir);
 	int FAT32_dir_empty(inode_s *dir);
